// src/app/models/inspection/inspection.ts
export interface IInspection {
  causa?: string;
  celInspector?: string;
  celLiquidador?: string;
  celularContratante?: string;
  celularsegurado?: string;
  ciudad?: string;
  ubicacion?: string;
  codCompania?: string;
  codInspector?: string;
  cod_inspeccion?: string;
  comuna?: string;
  corredor?: string;
  datos_siniestro?: string;
  direccion_siniestro?: string;
  emailContratante?: string;
  emailCorredor?: string;
  emailInspector?: string;
  emailLiquidador?: string;
  email_contacto?: string;
  emailasegurado?: string;
  fechaSiniestro?: string;
  fonoContratante?: string;
  fono_Contacto?: string;
  fonosegurado?: string;
  horaSiniestro?: string;
  id_accion?: string;
  id_tiposiniestro?: string;
  lugar?: string;
  nomAsegurado?: string;
  nomCompania?: string;
  nomContratante?: string;
  nomInspector?: string;
  nomLiquidador?: string;
  nomTipoSiniestro?: string;
  nombre_Contacto?: string;
  nro_carpeta?: string;
  numSiniestro?: string;
  rutAsegurado?: string;
  rutContratante?: string;
  rutInspector?: string;
  rut_Contacto?: string;
  telInspector?: string;
  telLiquidador?: string;
  valor_moneda?: string;
}

export class Inspection implements IInspection {
    causa?: string;
    celInspector?: string;
    ubicacion?: string;
    celLiquidador?: string;
    celularContratante?: string;
    celularsegurado?: string;
    ciudad?: string;
    codCompania?: string;
    codInspector?: string;
    cod_inspeccion?: string;
    comuna?: string;
    corredor?: string;
    datos_siniestro?: string;
    direccion_siniestro?: string;
    emailContratante?: string;
    emailCorredor?: string;
    emailInspector?: string;
    emailLiquidador?: string;
    email_contacto?: string;
    emailasegurado?: string;
    fechaSiniestro?: string;
    fonoContratante?: string;
    fono_Contacto?: string;
    fonosegurado?: string;
    horaSiniestro?: string;
    id_accion?: string;
    id_tiposiniestro?: string;
    lugar?: string;
    nomAsegurado?: string;
    nomCompania?: string;
    nomContratante?: string;
    nomInspector?: string;
    nomLiquidador?: string;
    nomTipoSiniestro?: string;
    nombre_Contacto?: string;
    nro_carpeta?: string;
    numSiniestro?: string;
    rutAsegurado?: string;
    rutContratante?: string;
    rutInspector?: string;
    rut_Contacto?: string;
    telInspector?: string;
    telLiquidador?: string;
    valor_moneda?: string;
    active?: number;
    status?: string;
    monto_Asegurado?:string;
    deducible?:string;
    moneda_Poliza?:string;
    sync?:number;
    completed?:number;
    pais?:string;
    updated_at?:string;
}